fix(#244): add missing workload type conversion
