+++ client/gefyra/api/utils.py
@@ -4,6 +4,26 @@ import time
 logger = logging.getLogger(__name__)
 
 
+def get_workload_type(workload_type_str: str):
+    POD = ["pod", "po", "pods"]
+    DEPLOYMENT = ["deploy", "deployment", "deployments"]
+    STATEFULSET = ["statefulset", "sts", "statefulsets"]
+    VALID_TYPES = POD + DEPLOYMENT + STATEFULSET
+
+    if workload_type_str not in VALID_TYPES:
+        raise RuntimeError(
+            f"Unknown workload type {workload_type_str}\n"
+            f'Valid workload types include: {", ".join(str(valid_type) for valid_type in VALID_TYPES)}'
+        )
+
+    if workload_type_str in POD:
+        return "pod"
+    elif workload_type_str in DEPLOYMENT:
+        return "deployment"
+    elif workload_type_str in STATEFULSET:
+        return "statefulset"
+
+
 def stopwatch(func):
     def wrapper(*args, **kwargs):
         tic = time.perf_counter()
